+++ src/assets/locales/en/en.js
@@ -3,6 +3,7 @@ import common from './common'
 import validations from '@/assets/locales/en/validations'
 import services from '@/assets/locales/en/services'
 import drivers from "@/assets/locales/en/drivers";
+import wp from "@/assets/locales/en/wp";
 
 export default {
   routes: {
@@ -17,5 +18,6 @@ export default {
   common: common,
   validations: validations,
   services: services,
-  drivers: drivers
+  drivers: drivers,
+  wp: wp
 }