+++ src/constants/Constants.ts
@@ -1,7 +1,8 @@
 export class Constants {
   static readonly DOC_TYPE_CC = 'CC'
   static readonly DOC_TYPE_CE = 'CE'
-  
+  static readonly DEFAULT_CLIENT = 'default-client'
+
   static readonly DOC_TYPES = [
     Constants.DOC_TYPE_CC,
     Constants.DOC_TYPE_CE