+++ src/types/ClientDiccionary.ts
@@ -1,0 +1,5 @@
+import {WpClient} from "@/types/WpClient";
+
+export interface ClientDictionary {
+    [id: string]: WpClient
+}