+++ tests/mocks/firebase/DocumentDataMock.ts
@@ -3,7 +3,7 @@ import ServiceMock from '../entities/ServiceMock'
 
 class DocumentDataMock implements DocumentData {
 	data = () => {
-		return ServiceMock
+		return new ServiceMock
 	}
 }