+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Pagination added with cursors of the consultation with firestore [#117](https://github.com/DevAlexandreCR/gorda-admin/pull/117)
 - added function to autocomplete if user does not exist [#118](https://github.com/DevAlexandreCR/gorda-admin/pull/118)
 
 # Release Notes for 1.2.x